make adjustments to Categories
