+++ src/components/Categories/categories.scss
@@ -1,0 +1,1 @@
+@import '../../styles/utils/variables';
