+++ src/styles/utils/_grid.scss
@@ -74,3 +74,15 @@
     }
   }
 }
+
+.section__category {
+  grid-column: span 4;
+  
+  @include on-tablet {
+    grid-column: span 12;
+  }
+
+  @include on-desktop {
+    grid-column: span 24;
+  }
+}
